requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
Unicorn
gunicorn
# ecdsa
# paramiko
# pycrypto
